Moved select project dialog
